+++ Meta/Generators/libweb_bindings/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (c) 2026-present, the Ladybird developers.
+#
+# SPDX-License-Identifier: BSD-2-Clause
